[QFEED-141] Refactor: Follow 엔티티 위치 변경
